amicsartromanic/api/requirements.txt
Flask==0.12
connexion>=1.0.97
Flask-Injector
fastavro
flask-sqlalchemy>=1.0.13
pymysql
flask-cors
Pillow
